backend/requirements.txt
# TrustChain Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
GitPython==3.1.40
py-algorand-sdk==2.6.0
pyteal==0.22.0
algokit-utils==2.2.0
firebase-admin==6.5.0
